+++ src/main/java/agh/edu/zeuspol/datastructures/types/attributes/Params.java
@@ -1,4 +1,4 @@
-package agh.edu.zeuspol.datastructures;
+package agh.edu.zeuspol.datastructures.types.attributes;
 
 import java.util.HashMap;
