+++ src/main/java/agh/edu/zeuspol/datastructures/types/attributes/RuleSubject.java
@@ -1,4 +1,4 @@
-package agh.edu.zeuspol.datastructures;
+package agh.edu.zeuspol.datastructures.types.attributes;
 
 public enum RuleSubject {
   RAM,
