+++ src/main/java/agh/edu/zeuspol/datastructures/types/attributes/UnitType.java
@@ -1,4 +1,4 @@
-package agh.edu.zeuspol.datastructures;
+package agh.edu.zeuspol.datastructures.types.attributes;
 
 public enum UnitType {
   NUMBER,
