+++ src/test/resources/drools/test_test.drl
@@ -1,5 +1,5 @@
 package drools;
-import agh.edu.zeuspol.datastructures.Rule;
+import agh.edu.zeuspol.datastructures.types.base.Rule;
 
 import io.github.hephaestusmetrics.model.metrics.Metric;
